backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python==4.9.0.80
ultralytics==8.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
numpy==1.26.3
